Complete Working Calculator V1
Making use of the eval function as it has the ability to calculate the entire string and further improving the efficiency of the program as I do not need the split each expression individually.

Using a replace function for multiplication and division expressions to allow the calculator to display 'x' and '÷' but still function normally.